requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
Pillow==10.0.0
openai==0.27.0
requests==2.31.0
drf-nested-routers==0.94.1
